refactor: minor code refactorings and scripts corrrection, add readme.md docs
