+++ jest.config.json
@@ -1,9 +1,9 @@
 {
   "preset": "ts-jest/presets/default-esm",
   "testEnvironment": "node",
-  "coveragePathIgnorePatterns": ["/node_modules/"],
   "extensionsToTreatAsEsm": [".ts"],
   "verbose": true,
+  "testPathIgnorePatterns": ["dist"],
   "moduleFileExtensions": ["ts", "js"],
   "transform": {
     "^.+\\.(ts|tsx)$": [
