+++ package.json
@@ -5,12 +5,13 @@
   "type": "module",
   "main": "dist/index.js",
   "scripts": {
-    "start1": "cross-env NODE_ENV=development PORT=process.env.PORT webpack serve --open --config webpack.dev.js",
-    "start2": "cross-env NODE_ENV=development PORT=process.env.PORT tsc && node dist/index.js",
-    "start:dev": "cross-env NODE_ENV=development PORT=process.env.PORT nodemon --watch src/**/*.ts -L --exec ts-node src/index.ts",
-    "start:dev2": "cross-env NODE_ENV=development PORT=process.env.PORT nodemon --watch src/**/*.ts -L",
-    "start:multi": "cross-env NODE_ENV=multi nodemon --watch src/**/*.ts -L --exec ts-node src/index.ts",
-    "build": "cross-env NODE_ENV=production PORT=process.env.PORT webpack --config webpack.prod.js",
+    "start:dev": "npm run lint && cross-env NODE_ENV=development nodemon --watch src/**/*.ts -L --exec ts-node --experimentalSpecifierResolution node src/index.ts",
+    "start:prod": "npm run lint && cross-env NODE_ENV=production tsc && node --experimental-specifier-resolution=node dist/src/index.js",
+    "start:multi": "npm run lint && cross-env NODE_ENV=multi nodemon --watch src/**/*.ts -L --exec ts-node --experimentalSpecifierResolution node src/index.ts",
+    "start1": "cross-env NODE_ENV=development webpack --config webpack.dev.js",
+    "start2": "cross-env NODE_ENV=development tsc && node dist/src/index.js",
+    "start:dev2": "cross-env NODE_ENV=development nodemon --watch src/**/*.ts -L",
+    "build": "cross-env NODE_ENV=production webpack --config webpack.prod.js",
     "lint": "eslint . --fix",
     "test": "npm run lint && cross-env NODE_OPTIONS=--experimental-vm-modules npx jest --detectOpenHandles --verbose"
   },
