+++ src/constants/http.constants.ts
@@ -1,3 +1,5 @@
+import 'dotenv/config';
+
 export const PORT = process.env.PORT || 4000;
 export const ENDPOINT_USERS = '/api/users';
 export const ENDPOINT_USERS_ID = `${ENDPOINT_USERS}/`;
@@ -8,6 +10,7 @@ export enum HttpStatusCode {
   NO_CONTENT = 204,
   BAD_REQUEST = 400,
   NOT_FOUND = 404,
+  INTERNAL_SERVER_ERROR = 500,
 }
 
 export enum HttpMethod {
