+++ src/data/user.data.ts
@@ -1,4 +1,4 @@
-import { User } from '../models/user.model.js';
+import { User } from '../models/user.model';
 
 export let usersStorage: User[] = [];
