+++ src/server.ts
@@ -1,64 +1,35 @@
 import * as http from 'node:http';
-import { PORT } from './constants/http.constants.js';
+import { PORT } from './constants/http.constants';
 import 'dotenv/config';
-import { UserController } from './controllers/user.controller.js';
-import {
-  HttpStatusCode,
-  HttpMethod,
-  ENDPOINT_USERS,
-  ENDPOINT_USERS_ID,
-} from './constants/http.constants.js';
 import cluster from 'node:cluster';
-import { User } from './models/user.model.js';
-import { usersStorage, modifyUsers } from './data/user.data.js';
-const workerPort = process.env.workerPort;
+import { User } from './models/user.model';
+import { usersStorage, modifyUsers } from './data/user.data';
+import { listenToRequests } from './controllers/request.listener';
 
-export async function listenToRequests(
-  req: http.IncomingMessage,
-  res: http.ServerResponse
-) {
-  if (req.url === ENDPOINT_USERS && req.method === HttpMethod.GET) {
-    await new UserController().getUsers(res);
-  } else if (
-    req.url?.startsWith(ENDPOINT_USERS_ID) &&
-    req.method === HttpMethod.GET
-  ) {
-    await new UserController().getUserById(req, res);
-  } else if (req.url === ENDPOINT_USERS && req.method === HttpMethod.POST) {
-    await new UserController().createUser(req, res);
-  } else if (
-    req.url?.startsWith(ENDPOINT_USERS_ID) &&
-    req.method === HttpMethod.PUT
-  ) {
-    await new UserController().updateUser(req, res);
-  } else if (
-    req.url?.startsWith(ENDPOINT_USERS_ID) &&
-    req.method === HttpMethod.DELETE
-  ) {
-    await new UserController().deleteUser(req, res);
-  } else {
-    res.writeHead(HttpStatusCode.NOT_FOUND, {
-      'Content-Type': 'application/json',
-    });
-    res.end(JSON.stringify({ message: 'Route not found' }));
-  }
-}
+const workerPort = process.env.workerPort;
+const environment = process.env.NODE_ENV;
+let portToListen = environment === 'multi' ? workerPort : PORT;
 
 export const server = http
   .createServer(async (req, res) => {
     await listenToRequests(req, res);
-    console.log(usersStorage, ' usersStorage from server1');
-    cluster.worker?.send({ users: usersStorage });
-    console.log(usersStorage, ' usersStorage from server2');
-    console.log(
-      `Worker ${String(cluster.worker?.id)} returned response on request`
-    );
+    if (environment === 'multi') {
+      cluster.worker?.send({ users: usersStorage });
+      console.log(
+        `Worker ${String(cluster.worker?.id)} returned response on request`
+      );
+    }
   })
-  .listen(workerPort || PORT, () => {
-    console.log(`Worker ${process.pid} started on ${String(workerPort)}`);
+  .listen(portToListen, () => {
+    const msg =
+      environment === 'multi'
+        ? `Worker ${process.pid} started on ${String(portToListen)}`
+        : `Server started on ${String(portToListen)}`;
+    console.log(msg);
   });
 
-process.on('message', (message: { users: User[] }) => {
-  console.log(message.users, 'from server');
-  modifyUsers(message.users);
-});
+if (environment === 'multi') {
+  process.on('message', (message: { users: User[] }) => {
+    modifyUsers(message.users);
+  });
+}
